requirements.txt
flask
torch
transformers
sentencepiece
python-dotenv